# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: 신촌쿵야 API
  version: 1.0.0
  description: 대학 단체 술자리 역경매 매칭 플랫폼의 P0 API 계약입니다. 인증 API와 메시지 API는 범위에서 제외합니다.
servers:
  - url: /
    description: 현재 접속 서버
...
paths:
  /stores:
    post:
      tags: [Store]
      summary: 가게 등록
      operationId: createStore
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreCreateReq'
            example:
              merchantId: 2
              name: 신촌포차
              region: SINCHON
              address: 서울 서대문구 연세로 1
              maxCapacity: 80
              openTime: '17:00:00'
              closeTime: '02:00:00'
              description: 단체 예약 가능한 포차
      responses:
        '201':
          description: 가게 등록 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreCreateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /stores/{id}/menus:
    post:
      tags: [Store]
      summary: 메뉴 일괄 등록
      operationId: createMenus
      parameters:
        - $ref: '#/components/parameters/StoreId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenuBulkCreateReq'
      responses:
        '201':
          description: 메뉴 일괄 등록 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuBulkCreateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  parameters:
    StoreId:
      name: id
      in: path
      required: true
      description: 가게 ID
      schema:
        type: integer
        format: int64
        minimum: 1
...
  responses:
    BadRequest:
      description: 요청값 또는 검증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
...
    NotFound:
      description: 요청한 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            isSuccess: false
            code: COMMON404_1
            message: 해당 리소스를 찾을 수 없습니다
            result: null
            error: null
...
    InternalServerError:
      description: 서버 내부 오류
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            isSuccess: false
            code: COMMON500_1
            message: 서버 내부 에러가 발생했습니다.
            result: null
            error: null
  schemas:
    StoreCreateReq:
      type: object
      required: [merchantId, name, region, address, maxCapacity]
      properties:
        merchantId: { type: integer, format: int64, minimum: 1, description: 인증 구현 전 임시 사장님 회원 ID }
        name: { type: string, minLength: 1, maxLength: 100 }
        region: { $ref: '#/components/schemas/Region' }
        address: { type: string, minLength: 1, maxLength: 255 }
        maxCapacity: { type: integer, minimum: 1 }
        openTime: { type: string, format: time, nullable: true, example: '17:00:00' }
        closeTime: { type: string, format: time, nullable: true, example: '02:00:00' }
        description: { type: string, maxLength: 500, nullable: true }
    MenuBulkCreateReq:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/MenuCreateReq'
      example:
        items:
          - name: 모둠전
            price: 25000
            category: MAIN
            servingSize: 4
          - name: 어묵탕
            price: 18000
            category: SOUP
            servingSize: 3
    MenuCreateReq:
      type: object
      required: [name, price, category]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        price: { type: integer, format: int64, minimum: 1 }
        category: { $ref: '#/components/schemas/MenuCategory' }
        servingSize: { type: integer, minimum: 1, nullable: true }
...
    StoreCreateResult:
      type: object
      required: [storeId]
      properties:
        storeId: { type: integer, format: int64, example: 1 }
    MenuBulkCreateResult:
      type: object
      required: [menuIds]
      properties:
        menuIds:
          type: array
          items: { type: integer, format: int64 }
          example: [1, 2]
...
    ErrorResponse:
      type: object
      required: [isSuccess, code, message]
      properties:
        isSuccess: { type: boolean, example: false }
        code: { type: string, example: COMMON404_1 }
        message: { type: string, example: 해당 리소스를 찾을 수 없습니다 }
        result: { nullable: true }
        error: { nullable: true }
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            error:
              type: object
              additionalProperties: { type: string }
              example:
                headcount: 인원은 10명 이상이어야 합니다.
    Region:
      type: string
      enum: [SINCHON, EWHA, HONGDAE]
    MenuCategory:
      type: string
      enum: [MAIN, FRIED, SOUP, DRY, SIDE, ETC]
...
    StoreCreateResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            result: { $ref: '#/components/schemas/StoreCreateResult' }
    MenuBulkCreateResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            result: { $ref: '#/components/schemas/MenuBulkCreateResult' }
...
    SuccessEnvelope:
      type: object
      required: [isSuccess, code, message]
      properties:
        isSuccess: { type: boolean, example: true }
        code: { type: string, example: COMMON200 }
        message: { type: string, example: 성공적으로 처리했습니다. }
        result: {}
        error: { nullable: true, example: null }